+++ src/Datos/Cliente.java
@@ -26,7 +26,7 @@ public class Cliente {
     private String nombre;
     private String apellido;
     private String domicilio;
-    private int telefono;
+    private long telefono;
     private boolean estado;
 
     public Cliente() {
@@ -70,11 +70,11 @@ public class Cliente {
         this.domicilio = domicilio;
     }
 
-    public int getTelefono() {
+    public long getTelefono() {
         return telefono;
     }
 
-    public void setTelefono(int telefono) {
+    public void setTelefono(long telefono) {
         this.telefono = telefono;
     }
 
@@ -87,7 +87,7 @@ public class Cliente {
     }
     
     
-public ResultSet consultaCliente(int telef) throws ClassNotFoundException{
+public ResultSet consultaCliente(long telef) throws ClassNotFoundException{
         try {
             Connection conex = Conexion.Cadena();            
             String ConsultaSQL = "SELECT * FROM cliente WHERE telefono = '" + telef + "'"; 
@@ -128,7 +128,7 @@ public int obtenerSiguienteId() throws ClassNotFoundException, SQLException{
             psPrepSencencias.setString(1, nombre);
             psPrepSencencias.setString(2, apellido);
             psPrepSencencias.setString(3, domicilio);
-            psPrepSencencias.setInt(4, telefono);
+            psPrepSencencias.setLong(4, telefono);
             psPrepSencencias.setBoolean(5, true);
             //ejecuto sentencia
             psPrepSencencias.executeUpdate();
@@ -171,7 +171,7 @@ public int obtenerSiguienteId() throws ClassNotFoundException, SQLException{
             psPrepSencencias.setString(1, nombre);
             psPrepSencencias.setString(2, apellido);
             psPrepSencencias.setString(3, domicilio);
-            psPrepSencencias.setInt(4, telefono);
+            psPrepSencencias.setLong(4, telefono);
             psPrepSencencias.setBoolean(5, true);
             psPrepSencencias.setInt(6, idCliente);
             //ejecuto sentencia
